Got the core functionality of download() working. The function now parses a completed table response and either returns the data to the terminal, or it writes to file and returns the data to the terminal.
Minor changes

Minor changes.

Minor changes.

+++ astroquery/cosmosim/core.py
@@ -1,6 +1,7 @@
 import requests
 import sys
 from lxml import etree
+import numpy as np
 
 # Astropy imports
 from astropy.table import Table
@@ -26,6 +27,8 @@ class CosmoSim(QueryWithLogin):
 
     QUERY_URL = COSMOSIM_SERVER()
     TIMEOUT = COSMOSIM_TIMEOUT()
+
+    cosmosim_databases = ('MDR1','MDPL','Bolshoi','BolshoiP')
 
     def __init__(self,username=None,password=None):
         self.username = username
@@ -113,15 +116,18 @@ class CosmoSim(QueryWithLogin):
         jobid : string
             The jobid of the sql query. If none provided
         """
-
+        
         if jobid is None:
             if hasattr(self,'current_job'):
                 jobid = self.current_job
             else:
-                print "Fix this"
+                try:
+                    jobid = self.current_job
+                except: 
+                    raise AttributeError
                 
         response = self.session.get(CosmoSim.QUERY_URL+'/{}'.format(jobid)+'/phase',auth=(self.username,self.password),data={'print':'b'})
-        print "Job {}: {}".format(self.current_job,response.content)
+        print "Job {}: {}".format(jobid,response.content)
         return response.content
 
     def check_all_jobs(self):
@@ -139,7 +145,7 @@ class CosmoSim(QueryWithLogin):
         root = etree.fromstring(checkalljobs.content)
         
         for iter in root:
-            if iter.find('{*}phase').text in ['COMPLETED','EXECUTING']:
+            if iter.find('{*}phase').text in ['COMPLETED','EXECUTING','ABORTED','ERROR']:
                 self.job_dict['{}'.format(iter.values()[1].split(CosmoSim.QUERY_URL+"/")[1])] = iter.find('{*}phase').text
             else:
                 self.job_dict['{}'.format(iter.values()[0])] = iter.find('{*}phase').text
@@ -185,14 +191,23 @@ class CosmoSim(QueryWithLogin):
             if hasattr(self,'current_job'):
                 if jobid == self.current_job:
                     del self.current_job
-        
-        result = self.session.delete(CosmoSim.QUERY_URL+"/{}".format(jobid),auth=(self.username,self.password),data={'follow':''})
+
+        if job_dict[jobid] in ['COMPLETED','ERROR','ABORTED']:
+            result = self.session.delete(CosmoSim.QUERY_URL+"/{}".format(jobid),auth=(self.username,self.password),data={'follow':''})
+        else:
+            print "Can only delete a job with phase: 'COMPLETED','ERROR',or 'ABORTED'."
         pdb.set_trace()
         if not result.ok:
             result.raise_for_status()
         print 'Deleted job: {}'.format(jobid)
         
         return result
+
+    def abort_job(self,jobid=None):
+        """
+        """
+
+        self.check_all_jobs()
 
     def delete_all_jobs(self):
 
@@ -206,8 +221,20 @@ class CosmoSim(QueryWithLogin):
 
         return 
 
-    def download(self,jobid=None,method=None):
+    def download(self,jobid=None,filename=None):
         """
+        A public function to download data from a job with COMPLETED phase.
+
+        Keyword Args
+        ----------
+        jobid :
+            Completed jobid to be downloaded
+        filename : string
+            If left blank, downloaded to the terminal. If specified, data is written out to file (directory can be included here).
+
+        Returns
+        -------
+        headers, data : list, list
         """
 
         if jobid is None:
@@ -216,15 +243,28 @@ class CosmoSim(QueryWithLogin):
             except:
                 raise
                 
-        
-        
         self.check_all_jobs()
         completed_job_responses = self.completed_job_info(jobid)
-        all_completed_jobs = [key for key in self.job_dict.keys() if self.job_dict[key] == 'COMPLETED']
-        pdb.set_trace()
+        root = etree.fromstring(completed_job_responses[0].content)
+        tableurl = [[list(c.attrib.values())[1] for c in e] for e in root.iter('{*}results') ][0][0]
+
+        # This is where the requestrequest.content parsing happens
+        raw_table_data = self.session.get(tableurl,auth=(self.username,self.password))
+        raw_headers = raw_table_data.content.split('\n')[0]
+        num_cols = len(raw_headers.split(','))
+        num_rows = len(raw_table_data.content.split('\n'))-2
+        headers = [raw_headers.split(',')[i].strip('"') for i in range(num_cols)]
+        raw_data = [raw_table_data.content.split('\n')[i+1].split(",") for i in range(num_rows)]
+        data = [map(eval,raw_data[i]) for i in range(num_rows)]
         
-        
-        
-        return
-    
-    
+        if filename is None:
+            return headers, data
+        else:
+            with open(filename, 'wb') as fh:
+                raw_table_data = self.session.get(tableurl,auth=(self.username,self.password),stream=True)
+                for block in raw_table_data.iter_content(1024):
+                    if not block:
+                        break
+                    fh.write(block)
+                print "Data written to file: {}".format(filename)
+            return headers, data
